pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tools.black]
line-length = 100

[tool.poetry]
name = "prisma-cloud-pipeline" 
version = "0.0.0" # replaced during build, see Makefile
description = "Export Prisma Cloud container findings to a CI pipeline, and identify un-triaged findings" 

license = "MIT"

authors = ["Michael Macnair <michael.macnair@thalesgroup.com>"]

readme = 'README.md'

repository = "https://github.com/thalescpl-io/prisma-cloud-pipeline"
homepage = "https://github.com/thalescpl-io/prisma-cloud-pipeline"

keywords = ['static-analysis', 'twistlock', 'devsecops', 'prisma-cloud']

[tool.poetry.dependencies]
python = "^3.6"
tabulate = "~0.8"
jq = "~1.1"
jsonschema = "~3.2"
requests = "~2.25"
pyyaml = "~5.3"

[tool.poetry.dev-dependencies]
mypy = ">=0.790"
pylint = ">=2.6.0"

[tool.poetry.scripts]
prisma-cloud-pipeline = 'prisma_cloud_pipeline.__main__:main'
